phased out dpop request handler
